Fixed json output
