+++ tests/test_configurations.py
@@ -6,7 +6,7 @@ import toml
 import subprocess
 
 from lib.lib_package_linter import *
-from lib.lib_package_linter import _print
+from lib.print import _print
 
 
 class Configurations(TestSuite):
