+++ tests/test_scripts.py
@@ -6,7 +6,7 @@ import subprocess
 import statistics
 
 from lib.lib_package_linter import *
-from lib.lib_package_linter import _print
+from lib.print import _print
 
 
 ##################################
